Move PointGroupType maps to const. Code review
